TripIt API Sucks, Part I.
Unfortunately we need to perform an OAuth v1 authentication flow so that
we can fetch the user's trips securely. TripIt's documentation and
`flyertools/tripit` are very lacking on what to do here, so we are going
to try and perform the OAuth flow ourselves and persist the user
credentials somewhere (probably DynamoDB in production and a local
MongoDB instance for testing).